Debug output for stream selection

+++ src/dash/DASHStream.cpp
@@ -239,7 +239,7 @@ bool DASHStream::select_stream(bool force, bool justInit, unsigned int repId)
     for (std::vector<DASHTree::Representation*>::const_iterator br(current_adp_->repesentations_.begin()), er(current_adp_->repesentations_.end()); br != er; ++br)
     {
       unsigned int dist;
-      if ((dist = abs((*br)->width_ * (*br)->height_ - width_ * height_)) < bestDist && (*br)->bandwidth_ <= bandwidth_
+      if ((dist = abs(static_cast<int>((*br)->width_ * (*br)->height_) - static_cast<int>(width_ * height_))) < bestDist && (*br)->bandwidth_ <= bandwidth_
         && (!new_rep || ((*br)->bandwidth_ > new_rep->bandwidth_)))
       {
         bestDist = dist;
